// libs/features/medical-board/src/lib/shared/enums/person-type-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum PersonTypeEnum {
  ContractedDoctor = 'Contracted Doctor',
  Nurse = 'Nurse',
  VisitingDoctor = 'Visiting Doctor',
  GosiDoctor = 'Gosi Doctor'
 }